python: make bindings/python/util.py symlink to bindings/python/oofem/util.py (experimental)
